spec installer to _metakernel module; run_as_main.
All classes are tightly coupled anyways and having everything together
avoids having to create a weird InstallerApp subclass on the fly.

Also drop pretense of import working in absence of IPython and Jupyter,
this was already not the case due to the import in metakernel.config.
