+++ metakernel/__init__.py
@@ -1,6 +1,5 @@
 from ._metakernel import (
-    MetaKernel, MetaKernelApp, IPythonKernel,
-    register_ipython_magics, get_metakernel)
+    MetaKernel, IPythonKernel, register_ipython_magics, get_metakernel)
 from . import pexpect
 from .replwrap import REPLWrapper, u
 from .process_metakernel import ProcessMetaKernel
