+++ metakernel_python/metakernel_python.py
@@ -69,5 +69,4 @@ class MetaKernelPython(MetaKernel):
 
 
 if __name__ == '__main__':
-    from metakernel import MetaKernelApp
-    MetaKernelApp.launch_instance(kernel_class=MetaKernelPython)
+    MetaKernelPython.run_as_main()
